+++ test/src/controlllers/$ScopeProvider.spec.js
@@ -1,11 +1,12 @@
 // System Modules
-import {expect} from        'chai';
-import simple, {mock} from          'simple-mock';
+import { expect } from          'chai';
+import simple, { mock } from    'simple-mock';
+
 
 // Angie Modules
-import $ScopeProvider, {
-    $scope
-} from                      '../../../src/controllers/$ScopeProvider';
+const TEST_ENV =                global.TEST_ENV || 'src',
+    $ScopeProvider =            require(`../../../${TEST_ENV}/controllers/$ScopeProvider`).default,
+    $scope =                    require(`../../../${TEST_ENV}/controllers/$ScopeProvider`).$scope;
 
 describe('$ScopeProvider', function() {
     let $$scope;