+++ test/src/factories/$CacheFactory.spec.js
@@ -1,8 +1,9 @@
 // System Modules
-import {expect} from    'chai';
+import { expect } from  'chai';
 
 // Angie Modules
-import $cacheFactory from '../../../src/factories/$CacheFactory';
+const TEST_ENV =        global.TEST_ENV || 'src',
+    $cacheFactory =     require(`../../../${TEST_ENV}/factories/$CacheFactory`);
 
 describe('$CacheFactory', function() {
     let cache;