+++ md/CHANGELOG.md
@@ -2,6 +2,17 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+#### [0.4.4] - 2015-10-19
+##### Added/Changed/Fixed
+- Refactored the `$ExceptionsProvider` to `$Exceptions` and properly placed it in the services directory
+- Added tests for `$Exceptions`
+- Added tests for the `Config` internal class
+- Added `$Log` as a registered module (Fix)
+- Added WallabyJS configuration
+- Add the ability to set X-Frame-Options header, as well as provisional security headers, in the response
+- Modified `.editorconfig` to include a global file configuration as well as js/css/html
+- Exposed the NodeJS webserver as the `$server` service.
+
 #### [0.4.3] - 2015-09-29
 ##### Changed
 - Modified README